# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server OA integrator.",
        "title": "Swagger OA Integrator API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/oa/{vendor}/AuthorizationService3rdParty/{facility}/{device}/{jobId}": {
            "post": {
                "description": "Creates new job and sends the required information as URI and \u003cjob\u003e element to 3rd party system.",
                "consumes": [
                    "application/xml"
                ],
                "produces": [
                    "application/xml"
                ],
                "tags": [
                    "oa"
                ],
                "summary": "S\u0026B creates new job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility",
                        "name": "facility",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device",
                        "name": "device",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Vendor",
                        "name": "vendor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/oa.JobWrapper"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/oa/{vendor}/AuthorizationService3rdParty/{facility}/{device}/{jobId}/cancel": {
            "put": {
                "description": "This request cancels a running job on the 3rd party side. The job is identified by its resource /facility/device/jobid",
                "consumes": [
                    "application/xml"
                ],
                "produces": [
                    "application/xml"
                ],
                "tags": [
                    "oa"
                ],
                "summary": "Cancels a running job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility",
                        "name": "facility",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device",
                        "name": "device",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Vendor",
                        "name": "vendor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/oa.CancelJobWrapper"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/oa/{vendor}/AuthorizationService3rdParty/{facility}/{device}/{jobId}/finalmessage": {
            "put": {
                "description": "This request sends the last message for a job. The job is identified by its resources /facility/device/jobid",
                "consumes": [
                    "application/xml"
                ],
                "produces": [
                    "application/xml"
                ],
                "tags": [
                    "oa"
                ],
                "summary": "Receive Final Message from S\u0026B",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility",
                        "name": "facility",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device",
                        "name": "device",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Vendor",
                        "name": "vendor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/oa.FinalMessageSBWrapper"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/oa/{vendor}/AuthorizationService3rdParty/{facility}/{device}/{jobId}/medialist": {
            "post": {
                "description": "Creates new media data for an existing job and sends the required information as a \u003cmediaData\u003e element to the 3rd party system.",
                "consumes": [
                    "application/xml"
                ],
                "produces": [
                    "application/xml"
                ],
                "tags": [
                    "oa"
                ],
                "summary": "Creates new media data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Facility",
                        "name": "facility",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device",
                        "name": "device",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Vendor",
                        "name": "vendor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/oa.MediaDataWrapper"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "oa.BusinessTransaction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "oa.CancelJobWrapper": {
            "type": "object",
            "properties": {
                "cancel": {
                    "type": "object",
                    "properties": {
                        "reason": {
                            "type": "object",
                            "properties": {
                                "cancelCode": {
                                    "type": "string"
                                },
                                "reasonText": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "oa.Customer": {
            "type": "object",
            "properties": {
                "customerGroup": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                }
            }
        },
        "oa.CustomerInformation": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/oa.Customer"
                }
            }
        },
        "oa.FinalMessageSB": {
            "type": "object",
            "properties": {
                "finalState": {
                    "type": "string"
                },
                "paymentMedia": {
                    "type": "string"
                }
            }
        },
        "oa.FinalMessageSBWrapper": {
            "type": "object",
            "properties": {
                "finalMessageSB": {
                    "$ref": "#/definitions/oa.FinalMessageSB"
                }
            }
        },
        "oa.Job": {
            "type": "object",
            "properties": {
                "businessTransaction": {
                    "$ref": "#/definitions/oa.BusinessTransaction"
                },
                "customerInformation": {
                    "$ref": "#/definitions/oa.CustomerInformation"
                },
                "jobId": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        }
                    }
                },
                "jobType": {
                    "type": "string"
                },
                "mediaDataList": {
                    "type": "object",
                    "properties": {
                        "identifier": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        },
                        "mediaType": {
                            "type": "string"
                        }
                    }
                },
                "paymentData": {
                    "$ref": "#/definitions/oa.PaymentData"
                },
                "providerInformation": {
                    "type": "object",
                    "properties": {
                        "provider": {
                            "type": "object",
                            "properties": {
                                "providerId": {
                                    "type": "string"
                                },
                                "providerName": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "timeAndPlace": {
                    "type": "object",
                    "properties": {
                        "computer": {
                            "type": "object",
                            "properties": {
                                "computerNumber": {
                                    "type": "string"
                                }
                            }
                        },
                        "device": {
                            "type": "object",
                            "properties": {
                                "deviceNumber": {
                                    "type": "string"
                                },
                                "deviceType": {
                                    "type": "string"
                                }
                            }
                        },
                        "facility": {
                            "type": "object",
                            "properties": {
                                "facilityNumber": {
                                    "type": "string"
                                }
                            }
                        },
                        "operator": {
                            "type": "object",
                            "properties": {
                                "operatorNumber": {
                                    "type": "string"
                                }
                            }
                        },
                        "transactionTimeStamp": {
                            "type": "object",
                            "properties": {
                                "timeStamp": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "oa.JobWrapper": {
            "type": "object",
            "properties": {
                "job": {
                    "$ref": "#/definitions/oa.Job"
                }
            }
        },
        "oa.MediaDataWrapper": {
            "type": "object",
            "properties": {
                "mediaData": {
                    "type": "object",
                    "properties": {
                        "hashValue": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string"
                                }
                            }
                        },
                        "mediaType": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "oa.OriginalAmount": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "vatRate": {
                    "type": "string"
                }
            }
        },
        "oa.PaymentData": {
            "type": "object",
            "properties": {
                "originalAmount": {
                    "$ref": "#/definitions/oa.OriginalAmount"
                },
                "remainingAmount": {
                    "type": "object",
                    "properties": {
                        "amount": {
                            "type": "string"
                        },
                        "text": {
                            "type": "string"
                        },
                        "vatRate": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}